requirements.txt
alembic==1.8.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
certifi==2022.6.15
cffi==1.15.1
click==8.1.3
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
pyotp==2.6.0
qrcode==7.3.1
SQLAlchemy==1.4.39
Werkzeug==2.1.2
zipp==3.8.0
